test: streamline Qwen3.5 cache regression
